Issue #43 - Developed unit tests for radian angle modulo by number operation; - Implemented radian angle modulo by number operation;
- Developed unit tests for degree angle modulo by number operation;
- Implemented degree angle modulo by number operation;

- Developed unit tests for angle modulo by number operation;
- Implemented angle modulo by number operation;

- Developed static typical angle values for all angle structures;

- Refactored test data precision;

+++ src/QuantitativeWorld/Angular/DegreeAngle.cs
@@ -86,6 +86,6 @@ namespace QuantitativeWorld.Angular
         AngleUnit ILinearQuantity<AngleUnit>.Unit => ((ILinearQuantity<AngleUnit>)this).BaseUnit;
 
         public static DegreeAngle FromAngle(Angle angle) =>
-            new DegreeAngle((double)angle.Convert(AngleUnit.Arcsecond).Value);
+            new DegreeAngle(angle.Convert(AngleUnit.Arcsecond).Value);
     }
 }
